+++ daemon/daemon_test.go
@@ -4,4 +4,5 @@ import "testing"
 
 func TestDaemon(t *testing.T) {
 	// TODO: NewInstance
+	// TODO: LoadConfig
 }